pyproject.toml
[tool.poetry]
name = "task-tracker"
version = "0.1.0"
description = ""
authors = ["Mark <vismanmark@yandex.ru>"]
readme = "README.md"

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
check_untyped_defs = true
exclude = 'venv|tests|task_tracker|manage.py|migrations'

[tool.black]
line-length = 119
exclude = '''
/(
    \.git
  | \.venv
  | \.mypy_cache
  | \.pytest_cache
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
line_length = 119
multi_line_output = 3

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.5"
djangorestframework = "^3.16.1"
pillow = "^11.3.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.1.1"
drf-yasg = "^1.21.10"
redis = "^6.4.0"
gunicorn = "^23.0.0"
django-filter = "^25.1"
djangorestframework-simplejwt = "^5.5.1"
django-cors-headers = "^4.7.0"


[tool.poetry.group.lint.dependencies]
mypy = "^1.17.1"
black = "^25.1.0"
flake8 = "^7.3.0"
isort = "^6.0.1"


[tool.poetry.group.test.dependencies]
coverage = "^7.10.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
